move slow unit test to integretion
Move test: test_calc_geom_impact_polys situated in:
climada/util/test/test_lines_polys_handler.py:289

to: climada/test/test_util.py

the test takes around 16 seconds
